requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
future==0.16.0
gspread==0.6.2
httplib2==0.10.3
idna==2.5
oauth2client==4.1.2
pyasn1==0.3.2
pyasn1-modules==0.0.11
python-telegram-bot==7.0.1
requests==2.18.3
rsa==3.4.2
six==1.10.0
urllib3==1.22
